requirements.txt
asgiref==3.9.1
Django==5.2.5
django-enumfield==3.1
djangorestframework==3.16.1
sqlparse==0.5.3
tzdata==2025.2
